+++ components/Navigation/BaseNavigation.vue
@@ -1,10 +1,10 @@
 <template>
-  <AppBar v-model:active-drawer="drawerVisible"/>
-  <Drawer v-model:active="drawerVisible"/>
+  <NavigationAppBar v-model:active-drawer="drawerVisible"/>
+  <NavigationDrawer v-model:active="drawerVisible"/>
 </template>
 
 <script setup lang="ts">
-  import Drawer from './Drawer.vue';
-  import AppBar from './AppBar.vue';
+  import NavigationAppBar from './NavigationAppBar.vue';
+  import NavigationDrawer from './NavigationDrawer.vue';
   const drawerVisible = ref(false);
 </script>
